requirements.txt
selenium
pandas
mysql-connector-python
matplotlib
seaborn
sklearn
scikit-learn
sqlalchemy
joblib
Flask 
Flask-RESTful